Move collector into experimental module
